feat(workouts): extract reusable stepper button component
